refactor: fix mocks paths and inject package version without vite
