+++ packages/experience-builder-sdk/src/core/componentRegistry.ts
@@ -1,11 +1,15 @@
 import { ComponentRegistration, ComponentDefinition } from '../types';
-import { OUTGOING_EVENTS, INTERNAL_EVENTS } from '../constants';
+import {
+  OUTGOING_EVENTS,
+  INTERNAL_EVENTS,
+  SDK_VERSION,
+  CONTENTFUL_CONTAINER_ID,
+  CONTENTFUL_SECTION_ID,
+} from '../constants';
 import { builtInStyles as builtInStyleDefinitions } from './definitions/variables';
-import { CONTENTFUL_CONTAINER_ID, CONTENTFUL_SECTION_ID } from '../constants';
 import { ContentfulContainer } from '../components/ContentfulContainer';
 import { containerDefinition } from './definitions/components';
 import { sendMessage } from '../communication/sendMessage';
-import { SDK_VERSION } from './constants';
 
 const cloneObject = <T>(targetObject: T): T => {
   if (typeof structuredClone !== 'undefined') {
