--- packages/experience-builder-sdk/src/core/constants.ts
@@ -1,3 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore -- this is replaced statically by vite with the env variable (becomes undefined if missing)
-export const SDK_VERSION = import.meta.env.VITE_SDK_VERSION;
