+++ packages/experience-builder-sdk/src/core/index.ts
@@ -1,3 +1,2 @@
 export { Flex } from '../components/Flex';
-export { createExperience } from '../utils/createExperience';
 export * as fetchers from './fetchers';
