+++ packages/experience-builder-sdk/src/core/preview/EntityStore.spec.ts
@@ -7,7 +7,8 @@ import { EntityStore } from './EntityStore';
 
 const locale = 'en-US';
 
-jest.mock('../core/constants', () => ({
+jest.mock('../../constants', () => ({
+  ...jest.requireActual('../../constants'),
   SDK_VERSION: '0.0.0-test',
   __esModule: true,
 }));
