+++ packages/experience-builder-sdk/src/hooks/useExperienceBuilder.test.ts
@@ -9,7 +9,8 @@ const clientMock = {
 
 const experienceTypeId = 'books';
 
-jest.mock('../core/constants', () => ({
+jest.mock('../constants', () => ({
+  ...jest.requireActual('../constants'),
   SDK_VERSION: '0.0.0-test',
   __esModule: true,
 }));
