+++ packages/experience-builder-sdk/src/hooks/useFetchExperience.test.ts
@@ -12,7 +12,8 @@ const slug = 'hello-world';
 
 let clientMock: ContentfulClientApi<undefined>;
 
-jest.mock('../core/constants', () => ({
+jest.mock('../constants', () => ({
+  ...jest.requireActual('../constants'),
   SDK_VERSION: '0.0.0-test',
   __esModule: true,
 }));
