+++ packages/experience-builder-sdk/src/hooks/useFetchExperience.ts
@@ -3,7 +3,7 @@ import { useCallback, useState } from 'react';
 import { EntityStore } from '../core/preview/EntityStore';
 import { fetchReferencedEntities, fetchExperienceEntry } from '../core/fetchers';
 import { Experience, ExternalSDKMode } from '../types';
-import { createExperience } from '../core';
+import { createExperience } from '../utils/createExperience';
 
 const errorMessagesWhileFetching = {
   experience: 'Failed to fetch experience',
