+++ packages/experience-builder-sdk/src/index.ts
@@ -23,4 +23,5 @@ export {
   CONTENTFUL_CONTAINER_ID,
   LATEST_SCHEMA_VERSION,
 } from './constants';
-export { createExperience, fetchers } from './core';
+export { fetchers } from './core';
+export { createExperience } from './utils/createExperience';
