+++ packages/experience-builder-sdk/src/sdkVersion.ts
@@ -1,0 +1,1 @@
+export const SDK_VERSION = '2.8.0';
