+++ packages/experience-builder-sdk/src/utils/createExperience.spec.ts
@@ -3,7 +3,8 @@ import { compositionEntry } from '../../test/__fixtures__/composition';
 import { assets, entries } from '../../test/__fixtures__/entities';
 import { createExperience } from './createExperience';
 
-jest.mock('../core/constants', () => ({
+jest.mock('../constants', () => ({
+  ...jest.requireActual('../constants'),
   SDK_VERSION: '0.0.0-test',
   __esModule: true,
 }));
